+++ src/main/java/org/opensearch/flowframework/transport/ReprovisionWorkflowTransportAction.java
@@ -270,10 +270,26 @@ public class ReprovisionWorkflowTransportAction extends HandledTransportAction<R
         ActionListener<WorkflowResponse> listener
     ) {
         try {
-            threadPool.executor(PROVISION_WORKFLOW_THREAD_POOL).execute(() -> { executeWorkflow(template, workflowSequence, workflowId); });
+            threadPool.executor(PROVISION_WORKFLOW_THREAD_POOL).execute(() -> {
+                updateTemplate(template, workflowId);
+                executeWorkflow(template, workflowSequence, workflowId);
+            });
         } catch (Exception exception) {
             listener.onFailure(new FlowFrameworkException("Failed to execute workflow " + workflowId, ExceptionsHelper.status(exception)));
         }
+    }
+
+    /**
+     * Replace template document
+     * @param template The template to store after reprovisioning completes successfully
+     * @param workflowId The workflowId associated with the workflow that is executing
+     */
+    private void updateTemplate(Template template, String workflowId) {
+        flowFrameworkIndicesHandler.updateTemplateInGlobalContext(workflowId, template, ActionListener.wrap(templateResponse -> {
+            logger.info("Updated template for {}", workflowId);
+        }, exception -> { logger.error("Failed to update use case template for {}", workflowId, exception); }),
+            true  // ignores NOT_STARTED state if request is to reprovision
+        );
     }
 
     /**
@@ -289,8 +305,9 @@ public class ReprovisionWorkflowTransportAction extends HandledTransportAction<R
             for (ProcessNode processNode : workflowSequence) {
                 List<ProcessNode> predecessors = processNode.predecessors();
                 logger.info(
-                    "Queueing process [{}].{}",
+                    "Queueing Process [{} (type: {})].{}",
                     processNode.id(),
+                    processNode.workflowStep().getName(),
                     predecessors.isEmpty()
                         ? " Can start immediately!"
                         : String.format(
@@ -321,18 +338,6 @@ public class ReprovisionWorkflowTransportAction extends HandledTransportAction<R
 
                     logger.info("updated workflow {} state to {}", workflowId, State.COMPLETED);
 
-                    // Replace template document
-                    flowFrameworkIndicesHandler.updateTemplateInGlobalContext(
-                        workflowId,
-                        template,
-                        ActionListener.wrap(templateResponse -> {
-                            logger.info("Updated template for {}", workflowId, State.COMPLETED);
-                        }, exception -> {
-                            String errorMessage = "Failed to update use case template for " + workflowId;
-                            logger.error(errorMessage, exception);
-                        }),
-                        true  // ignores NOT_STARTED state if request is to reprovision
-                    );
                 }, exception -> { logger.error("Failed to update workflow state for workflow {}", workflowId, exception); })
             );
         } catch (Exception ex) {